Further work on GUI
Now there is a double click handler on the main table, from there the
rest of the mask is populated with information about a person. I figured
I should store the Oid with the tablerow elements to access the full
record of a person in a new transaction when needed.

I also made the declaration of all those names of GUI elements more
compact, its easier to read this way.